requirements.txt
# Production dependencies
# Generated: 2025-10-15T23:27:34.286837

faster-whisper==1.0.3
geocoder==1.38.1
geopy==2.4.1
groq==0.32.0
httpx==0.28.1
mlx-whisper==0.4.3
numpy==1.26.0
openai==2.3.0
openwakeword==0.5.1
python-dotenv==1.0.0
pytz==2025.2
pyyaml==6.0.1
sounddevice==0.4.6
structlog==24.1.0
timezonefinder==8.1.0
webrtcvad-wheels==2.0.14
